requirements.txt
# Dashboard Framework
streamlit
watchdog

# Data Processing
pandas
numpy
pyarrow
openpyxl
et_xmlfile

# Database
pymysql
sqlalchemy

# Visualization
plotly
altair
pydeck
pillow

# AI & Google
google-generativeai
google-auth
google-api-python-client

# Utilities
python-dotenv
requests
python-dateutil
pytz
streamlit-authenticator
pyyaml